feature flagged all languages, improved docs on language crate
