+++ crates/ast-engine/src/node.rs
@@ -19,7 +19,7 @@
 //! ```rust,no_run
 //! # use thread_ast_engine::Language;
 //! # use thread_ast_engine::tree_sitter::LanguageExt;
-//! # use thread_ast_engine::matcher::MatcherExt;
+//! # use thread_ast_engine::MatcherExt;
 //! let ast = Language::Tsx.ast_grep("function foo() { return 42; }");
 //! let root_node = ast.root();
 //!
@@ -38,6 +38,7 @@ use crate::Doc;
 use crate::Language;
 #[cfg(feature = "matching")]
 use crate::matcher::{Matcher, MatcherExt, NodeMatch};
+#[cfg(feature = "matching")]
 use crate::replacer::Replacer;
 use crate::source::{Content, Edit as E, SgNode};
 
@@ -78,14 +79,16 @@ pub struct Position {
 }
 
 impl Position {
-    #[must_use] pub const fn new(line: usize, byte_column: usize, byte_offset: usize) -> Self {
+    #[must_use]
+    pub const fn new(line: usize, byte_column: usize, byte_offset: usize) -> Self {
         Self {
             line,
             byte_column,
             byte_offset,
         }
     }
-    #[must_use] pub const fn line(&self) -> usize {
+    #[must_use]
+    pub const fn line(&self) -> usize {
         self.line
     }
     /// Returns the column in terms of characters.
@@ -94,7 +97,8 @@ impl Position {
         let source = node.get_doc().get_source();
         source.get_char_column(self.byte_column, self.byte_offset)
     }
-    #[must_use] pub const fn byte_point(&self) -> (usize, usize) {
+    #[must_use]
+    pub const fn byte_point(&self) -> (usize, usize) {
         (self.line, self.byte_column)
     }
 }
@@ -114,7 +118,7 @@ impl Position {
 /// ```rust,no_run
 /// # use thread_ast_engine::Language;
 /// # use thread_ast_engine::tree_sitter::LanguageExt;
-/// # use thread_ast_engine::matcher::MatcherExt;
+/// # use thread_ast_engine::MatcherExt;
 /// let mut ast = Language::Tsx.ast_grep("let x = 42;");
 /// let root_node = ast.root();
 ///
@@ -149,6 +153,7 @@ impl<D: Doc> Root<D> {
         Ok(self)
     }
 
+    #[cfg(feature = "matching")]
     pub fn replace<M: Matcher, R: Replacer<D>>(
         &mut self,
         pattern: M,
@@ -165,7 +170,7 @@ impl<D: Doc> Root<D> {
     }
 
     /// Adopt the `tree_sitter` as the descendant of the root and return the wrapped sg Node.
-    /// It assumes `inner` is the under the root and will panic at dev build if wrong node is used.
+    /// It assumes `inner` is under the root and will panic at dev build if wrong node is used.
     pub fn adopt<'r>(&'r self, inner: D::Node<'r>) -> Node<'r, D> {
         debug_assert!(self.check_lineage(&inner));
         Node { inner, root: self }
@@ -301,6 +306,7 @@ impl<'r, D: Doc> Node<'r, D> {
 /**
  * Corresponds to inside/has/precedes/follows
  */
+#[cfg(feature = "matching")]
 impl<D: Doc> Node<'_, D> {
     pub fn matches<M: Matcher>(&self, m: M) -> bool {
         m.match_node(self.clone()).is_some()
@@ -426,11 +432,11 @@ impl<'r, D: Doc> Node<'r, D> {
         })
     }
 
-    #[must_use]
+    #[cfg(feature = "matching")]
     pub fn find<M: Matcher>(&self, pat: M) -> Option<NodeMatch<'r, D>> {
         pat.find_node(self.clone())
     }
-
+    #[cfg(feature = "matching")]
     pub fn find_all<'s, M: Matcher + 's>(
         &'s self,
         pat: M,
@@ -449,6 +455,7 @@ impl<'r, D: Doc> Node<'r, D> {
 
 /// Tree manipulation API
 impl<D: Doc> Node<'_, D> {
+    #[cfg(feature = "matching")]
     pub fn replace<M: Matcher, R: Replacer<D>>(&self, matcher: M, replacer: R) -> Option<Edit<D>> {
         let matched = matcher.find_node(self.clone())?;
         let edit = matched.make_edit(&matcher, &replacer);
