+++ crates/language/src/lib.rs
@@ -4,38 +4,87 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later AND MIT
 
-//! This module defines the supported programming languages for ast-grep.
+//! Language definitions and tree-sitter parsers for Thread AST analysis.
 //!
-//! It provides a set of customized languages with expando_char / pre_process_pattern,
-//! and a set of stub languages without preprocessing.
-//! A rule of thumb: if your language does not accept identifiers like `$VAR`.
-//! You need use `impl_lang_expando!` macro and a standalone file for testing.
-//! Otherwise, you can define it as a stub language using `impl_lang!`.
-//! To see the full list of languages, visit `<https://ast-grep.github.io/reference/languages.html>`
+//! Provides unified language support through consistent [`Language`] and [`LanguageExt`] traits
+//! across 24+ programming languages. Each language can be feature-gated individually or included
+//! in groups.
+//!
+//! ## Language Categories
+//!
+//! ### Standard Languages
+//! Languages that accept `$` as a valid identifier character and use default pattern processing:
+//! - [`Bash`], [`Java`], [`JavaScript`], [`Json`], [`Lua`], [`Scala`], [`TypeScript`], [`Tsx`], [`Yaml`]
+//!
+//! ### Custom Pattern Languages
+//! Languages requiring special metavariable handling with custom expando characters:
+//! - [`C`] (`µ`), [`Cpp`] (`µ`), [`CSharp`] (`µ`), [`Css`] (`_`), [`Elixir`] (`µ`)
+//! - [`Go`] (`µ`), [`Haskell`] (`µ`), [`Html`] (`z`), [`Kotlin`] (`µ`), [`Php`] (`µ`)
+//! - [`Python`] (`µ`), [`Ruby`] (`µ`), [`Rust`] (`µ`), [`Swift`] (`µ`)
+//!
+//! ## Usage
+//!
+//! ```rust
+//! use thread_language::{SupportLang, Rust};
+//! use thread_ast_engine::{Language, LanguageExt};
+//!
+//! // Runtime language selection
+//! let lang = SupportLang::from_path("main.rs").unwrap();
+//! let tree = lang.ast_grep("fn main() {}");
+//!
+//! // Compile-time language selection
+//! let rust = Rust;
+//! let tree = rust.ast_grep("fn main() {}");
+//! ```
+//!
+//! ## Implementation Details
+//!
+//! Languages are implemented using two macros:
+//! - [`impl_lang!`] - Standard languages accepting `$` in identifiers
+//! - [`impl_lang_expando!`] - Languages requiring custom expando characters for metavariables
 pub mod parsers;
 
+#[cfg(feature = "bash")]
 mod bash;
+#[cfg(feature = "cpp")]
 mod cpp;
+#[cfg(feature = "csharp")]
 mod csharp;
+#[cfg(feature = "css")]
 mod css;
+#[cfg(feature = "elixir")]
 mod elixir;
+#[cfg(feature = "go")]
 mod go;
+#[cfg(feature = "haskell")]
 mod haskell;
+#[cfg(feature = "html")]
 mod html;
+#[cfg(feature = "json")]
 mod json;
+#[cfg(feature = "kotlin")]
 mod kotlin;
+#[cfg(feature = "lua")]
 mod lua;
+#[cfg(feature = "php")]
 mod php;
 #[cfg(feature = "profiling")]
 pub mod profiling;
+#[cfg(feature = "python")]
 mod python;
+#[cfg(feature = "ruby")]
 mod ruby;
+#[cfg(feature = "rust")]
 mod rust;
+#[cfg(feature = "scala")]
 mod scala;
+#[cfg(feature = "swift")]
 mod swift;
+#[cfg(feature = "yaml")]
 mod yaml;
-
+#[cfg(feature = "html")]
 pub use html::Html;
+#[cfg(feature = "matching")]
 use thread_ast_engine::{Pattern, PatternBuilder, PatternError};
 
 use ignore::types::{Types, TypesBuilder};
@@ -46,15 +95,33 @@ use std::fmt;
 use std::fmt::{Display, Formatter};
 use std::path::Path;
 use std::str::FromStr;
+#[cfg(feature = "matching")]
 use thread_ast_engine::Node;
 use thread_ast_engine::meta_var::MetaVariable;
-use thread_ast_engine::tree_sitter::{StrDoc, TSLanguage, TSRange};
+use thread_ast_engine::tree_sitter::TSLanguage;
+#[cfg(feature = "matching")]
+use thread_ast_engine::tree_sitter::{StrDoc, TSRange};
+#[cfg(feature = "matching")]
 use thread_utils::RapidMap;
 
 pub use thread_ast_engine::language::Language;
 pub use thread_ast_engine::tree_sitter::LanguageExt;
 
-/// this macro implements bare-bone methods for a language
+/// Implements standard [`Language`] and [`LanguageExt`] traits for languages that accept `$` in identifiers.
+///
+/// Used for languages like JavaScript, Python, and Rust where `$` can appear in variable names
+/// and doesn't require special preprocessing for metavariables.
+///
+/// # Parameters
+/// - `$lang` - The language struct name (e.g., `JavaScript`)
+/// - `$func` - The parser function name from [`parsers`] module (e.g., `language_javascript`)
+///
+/// # Generated Implementation
+/// Creates a zero-sized struct with [`Language`] and [`LanguageExt`] implementations that:
+/// - Map node kinds and field names to tree-sitter IDs
+/// - Build patterns using the language's parser
+/// - Use default metavariable processing (no expando character substitution)
+#[allow(unused_macros)]
 macro_rules! impl_lang {
     ($lang: ident, $func: ident) => {
         #[derive(Clone, Copy, Debug)]
@@ -69,6 +136,7 @@ macro_rules! impl_lang {
                     .field_id_for_name(field)
                     .map(|f| f.get())
             }
+            #[cfg(feature = "matching")]
             fn build_pattern(&self, builder: &PatternBuilder) -> Result<Pattern, PatternError> {
                 builder.build(|src| StrDoc::try_new(src, self.clone()))
             }
@@ -81,6 +149,31 @@ macro_rules! impl_lang {
     };
 }
 
+/// Preprocesses pattern strings by replacing `$` with the language's expando character.
+///
+/// Languages that don't accept `$` in identifiers need metavariables like `$VAR` converted
+/// to use a different character. This function efficiently replaces `$` symbols that precede
+/// uppercase letters, underscores, or appear in triple sequences (`$$$`).
+///
+/// # Parameters
+/// - `expando` - The character to replace `$` with (e.g., `µ` for most languages, `_` for CSS)
+/// - `query` - The pattern string containing `$` metavariables
+///
+/// # Returns
+/// - `Cow::Borrowed` if no replacement is needed (fast path)
+/// - `Cow::Owned` if replacement occurred
+///
+/// # Examples
+/// ```rust
+/// # use thread_language::pre_process_pattern;
+/// // Python doesn't accept $ in identifiers, so use µ
+/// let result = pre_process_pattern('µ', "def $FUNC($ARG): pass");
+/// assert_eq!(result, "def µFUNC(µARG): pass");
+///
+/// // No change needed
+/// let result = pre_process_pattern('µ', "def hello(): pass");
+/// assert_eq!(result, "def hello(): pass");
+/// ```
 fn pre_process_pattern(expando: char, query: &str) -> std::borrow::Cow<'_, str> {
     // Fast path: check if any processing is needed
     let has_dollar = query.as_bytes().contains(&b'$');
@@ -141,8 +234,35 @@ fn pre_process_pattern(expando: char, query: &str) -> std::borrow::Cow<'_, str> 
     std::borrow::Cow::Owned(ret)
 }
 
-/// this macro will implement expando_char and pre_process_pattern
-/// use this if your language does not accept $ as valid identifier char
+/// Implements [`Language`] and [`LanguageExt`] traits for languages requiring custom expando characters.
+///
+/// Used for languages that don't accept `$` in identifiers and need metavariables like `$VAR`
+/// converted to use a different character (e.g., `µVAR`, `_VAR`).
+///
+/// # Parameters
+/// - `$lang` - The language struct name (e.g., `Python`)
+/// - `$func` - The parser function name from [`parsers`] module (e.g., `language_python`)
+/// - `$char` - The expando character to use instead of `$` (e.g., `'µ'`)
+///
+/// # Generated Implementation
+/// Creates a zero-sized struct with [`Language`] and [`LanguageExt`] implementations that:
+/// - Map node kinds and field names to tree-sitter IDs
+/// - Build patterns using the language's parser
+/// - Preprocess patterns by replacing `$` with the expando character
+/// - Provide the expando character via [`Language::expando_char`]
+///
+/// # Examples
+/// ```rust
+/// # use thread_language::Python;
+/// # use thread_ast_engine::Language;
+/// let python = Python;
+/// assert_eq!(python.expando_char(), 'µ');
+///
+/// // Pattern gets automatically preprocessed
+/// let pattern = "def $FUNC($ARG): pass";
+/// let processed = python.pre_process_pattern(pattern);
+/// assert_eq!(processed, "def µFUNC(µARG): pass");
+/// ```
 macro_rules! impl_lang_expando {
     ($lang: ident, $func: ident, $char: expr) => {
         #[derive(Clone, Copy, Debug)]
@@ -163,6 +283,7 @@ macro_rules! impl_lang_expando {
             fn pre_process_pattern<'q>(&self, query: &'q str) -> std::borrow::Cow<'q, str> {
                 pre_process_pattern(self.expando_char(), query)
             }
+            #[cfg(feature = "matching")]
             fn build_pattern(&self, builder: &PatternBuilder) -> Result<Pattern, PatternError> {
                 builder.build(|src| StrDoc::try_new(src, self.clone()))
             }
@@ -216,92 +337,180 @@ macro_rules! impl_alias {
 /// Generates as convenience conversions between the lang types
 /// and `SupportedType`.
 macro_rules! impl_aliases {
-  ($($lang:ident => $as:expr),* $(,)?) => {
-    $(impl_alias!($lang => $as);)*
+  ($($lang:ident, $feature:literal => $as:expr),* $(,)?) => {
+    $(#[cfg(feature = $feature)]
+      impl_alias!($lang => $as);
+    )*
     const fn alias(lang: SupportLang) -> &'static [&'static str] {
       match lang {
-        $(SupportLang::$lang => $lang::ALIAS),*
+        $(
+          #[cfg(feature = $feature)]
+          SupportLang::$lang => $lang::ALIAS,
+        )*
       }
     }
   };
 }
-
 /* Customized Language with expando_char / pre_process_pattern */
+
 // https://en.cppreference.com/w/cpp/language/identifiers
 // Due to some issues in the tree-sitter parser, it is not possible to use
 // unicode literals in identifiers for C/C++ parsers
+#[cfg(feature = "c")]
 impl_lang_expando!(C, language_c, 'µ');
+#[cfg(feature = "cpp")]
 impl_lang_expando!(Cpp, language_cpp, 'µ');
+
 // https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/language-specification/lexical-structure#643-identifiers
 // all letter number is accepted
 // https://www.compart.com/en/unicode/category/Nl
+#[cfg(feature = "csharp")]
 impl_lang_expando!(CSharp, language_c_sharp, 'µ');
+
 // https://www.w3.org/TR/CSS21/grammar.html#scanner
+#[cfg(feature = "css")]
 impl_lang_expando!(Css, language_css, '_');
+
 // https://github.com/elixir-lang/tree-sitter-elixir/blob/a2861e88a730287a60c11ea9299c033c7d076e30/grammar.js#L245
+#[cfg(feature = "elixir")]
 impl_lang_expando!(Elixir, language_elixir, 'µ');
+
 // we can use any Unicode code point categorized as "Letter"
 // https://go.dev/ref/spec#letter
+#[cfg(feature = "go")]
 impl_lang_expando!(Go, language_go, 'µ');
+
 // GHC supports Unicode syntax per
 // https://ghc.gitlab.haskell.org/ghc/doc/users_guide/exts/unicode_syntax.html
 // and the tree-sitter-haskell grammar parses it too.
+#[cfg(feature = "haskell")]
 impl_lang_expando!(Haskell, language_haskell, 'µ');
+
 // https://github.com/fwcd/tree-sitter-kotlin/pull/93
+#[cfg(feature = "kotlin")]
 impl_lang_expando!(Kotlin, language_kotlin, 'µ');
+
 // PHP accepts unicode to be used as some name not var name though
+#[cfg(feature = "php")]
 impl_lang_expando!(Php, language_php, 'µ');
+
 // we can use any char in unicode range [:XID_Start:]
 // https://docs.python.org/3/reference/lexical_analysis.html#identifiers
 // see also [PEP 3131](https://peps.python.org/pep-3131/) for further details.
+#[cfg(feature = "python")]
 impl_lang_expando!(Python, language_python, 'µ');
+
 // https://github.com/tree-sitter/tree-sitter-ruby/blob/f257f3f57833d584050336921773738a3fd8ca22/grammar.js#L30C26-L30C78
+#[cfg(feature = "ruby")]
 impl_lang_expando!(Ruby, language_ruby, 'µ');
+
 // we can use any char in unicode range [:XID_Start:]
 // https://doc.rust-lang.org/reference/identifiers.html
+#[cfg(feature = "rust")]
 impl_lang_expando!(Rust, language_rust, 'µ');
+
 //https://docs.swift.org/swift-book/documentation/the-swift-programming-language/lexicalstructure/#Identifiers
+#[cfg(feature = "swift")]
 impl_lang_expando!(Swift, language_swift, 'µ');
 
 // Stub Language without preprocessing
 // Language Name, tree-sitter-name, alias, extension
+#[cfg(feature = "bash")]
 impl_lang!(Bash, language_bash);
+#[cfg(feature = "java")]
 impl_lang!(Java, language_java);
+#[cfg(feature = "javascript")]
 impl_lang!(JavaScript, language_javascript);
+#[cfg(feature = "json")]
 impl_lang!(Json, language_json);
+#[cfg(feature = "lua")]
 impl_lang!(Lua, language_lua);
+#[cfg(feature = "scala")]
 impl_lang!(Scala, language_scala);
+#[cfg(feature = "tsx")]
 impl_lang!(Tsx, language_tsx);
+#[cfg(feature = "typescript")]
 impl_lang!(TypeScript, language_typescript);
+#[cfg(feature = "yaml")]
 impl_lang!(Yaml, language_yaml);
 // See ripgrep for extensions
 // https://github.com/BurntSushi/ripgrep/blob/master/crates/ignore/src/default_types.rs
 
-/// Represents all built-in languages.
+/// Runtime language selection enum supporting all built-in languages.
+///
+/// Provides a unified interface for working with any supported language at runtime.
+/// Each variant corresponds to a specific programming language implementation.
+///
+/// # Language Detection
+/// ```rust,ignore
+/// use thread_language::SupportLang;
+/// use std::path::Path;
+///
+/// // Detect from file extension
+/// let lang = SupportLang::from_path("main.rs").unwrap();
+/// assert_eq!(lang, SupportLang::Rust);
+///
+/// // Parse from string
+/// let lang: SupportLang = "javascript".parse().unwrap();
+/// assert_eq!(lang, SupportLang::JavaScript);
+/// ```
+///
+/// # Usage with AST Analysis
+/// ```rust
+/// use thread_language::SupportLang;
+/// use thread_ast_engine::{Language, LanguageExt};
+///
+/// let lang = SupportLang::Rust;
+/// let tree = lang.ast_grep("fn main() {}");
+/// let pattern = lang.build_pattern(&pattern_builder).unwrap();
+/// ```
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Hash)]
 pub enum SupportLang {
+    #[cfg(feature = "bash")]
     Bash,
+    #[cfg(feature = "c")]
     C,
+    #[cfg(feature = "cpp")]
     Cpp,
+    #[cfg(feature = "csharp")]
     CSharp,
+    #[cfg(feature = "css")]
     Css,
+    #[cfg(feature = "go")]
     Go,
+    #[cfg(feature = "elixir")]
     Elixir,
+    #[cfg(feature = "haskell")]
     Haskell,
+    #[cfg(feature = "html")]
     Html,
+    #[cfg(feature = "java")]
     Java,
+    #[cfg(feature = "javascript")]
     JavaScript,
+    #[cfg(feature = "json")]
     Json,
+    #[cfg(feature = "kotlin")]
     Kotlin,
+    #[cfg(feature = "lua")]
     Lua,
+    #[cfg(feature = "php")]
     Php,
+    #[cfg(feature = "python")]
     Python,
+    #[cfg(feature = "ruby")]
     Ruby,
+    #[cfg(feature = "rust")]
     Rust,
+    #[cfg(feature = "scala")]
     Scala,
+    #[cfg(feature = "swift")]
     Swift,
+    #[cfg(feature = "tsx")]
     Tsx,
+    #[cfg(feature = "typescript")]
     TypeScript,
+    #[cfg(feature = "yaml")]
     Yaml,
 }
 
@@ -309,8 +518,52 @@ impl SupportLang {
     pub const fn all_langs() -> &'static [SupportLang] {
         use SupportLang::*;
         &[
-            Bash, C, Cpp, CSharp, Css, Elixir, Go, Haskell, Html, Java, JavaScript, Json, Kotlin,
-            Lua, Php, Python, Ruby, Rust, Scala, Swift, Tsx, TypeScript, Yaml,
+            #[cfg(feature = "bash")]
+            Bash,
+            #[cfg(feature = "c")]
+            C,
+            #[cfg(feature = "cpp")]
+            Cpp,
+            #[cfg(feature = "csharp")]
+            CSharp,
+            #[cfg(feature = "css")]
+            Css,
+            #[cfg(feature = "elixir")]
+            Elixir,
+            #[cfg(feature = "go")]
+            Go,
+            #[cfg(feature = "haskell")]
+            Haskell,
+            #[cfg(feature = "html")]
+            Html,
+            #[cfg(feature = "java")]
+            Java,
+            #[cfg(feature = "javascript")]
+            JavaScript,
+            #[cfg(feature = "json")]
+            Json,
+            #[cfg(feature = "kotlin")]
+            Kotlin,
+            #[cfg(feature = "lua")]
+            Lua,
+            #[cfg(feature = "php")]
+            Php,
+            #[cfg(feature = "python")]
+            Python,
+            #[cfg(feature = "ruby")]
+            Ruby,
+            #[cfg(feature = "rust")]
+            Rust,
+            #[cfg(feature = "scala")]
+            Scala,
+            #[cfg(feature = "swift")]
+            Swift,
+            #[cfg(feature = "tsx")]
+            Tsx,
+            #[cfg(feature = "typescript")]
+            TypeScript,
+            #[cfg(feature = "yaml")]
+            Yaml,
         ]
     }
 
@@ -328,6 +581,7 @@ impl fmt::Display for SupportLang {
 #[derive(Debug)]
 pub enum SupportLangErr {
     LanguageNotSupported(String),
+    LanguageNotEnabled(String)
 }
 
 impl Display for SupportLangErr {
@@ -335,6 +589,7 @@ impl Display for SupportLangErr {
         use SupportLangErr::*;
         match self {
             LanguageNotSupported(lang) => write!(f, "{lang} is not supported!"),
+            LanguageNotEnabled(lang) => write!(f, "{lang} is available but not enabled. You need to enable the feature flag for this language.")
         }
     }
 }
@@ -390,29 +645,29 @@ impl Visitor<'_> for AliasVisitor {
 }
 
 impl_aliases! {
-  Bash => &["bash"],
-  C => &["c"],
-  Cpp => &["cc", "c++", "cpp", "cxx"],
-  CSharp => &["cs", "csharp"],
-  Css => &["css"],
-  Elixir => &["ex", "elixir"],
-  Go => &["go", "golang"],
-  Haskell => &["hs", "haskell"],
-  Html => &["html"],
-  Java => &["java"],
-  JavaScript => &["javascript", "js", "jsx"],
-  Json => &["json"],
-  Kotlin => &["kotlin", "kt"],
-  Lua => &["lua"],
-  Php => &["php"],
-  Python => &["py", "python"],
-  Ruby => &["rb", "ruby"],
-  Rust => &["rs", "rust"],
-  Scala => &["scala"],
-  Swift => &["swift"],
-  TypeScript => &["ts", "typescript"],
-  Tsx => &["tsx"],
-  Yaml => &["yaml", "yml"],
+  Bash, "bash" => &["bash"],
+  C, "c" => &["c"],
+  Cpp, "cpp" => &["cc", "c++", "cpp", "cxx"],
+  CSharp, "csharp" => &["cs", "csharp"],
+  Css, "css" => &["css"],
+  Elixir, "elixir" => &["ex", "elixir"],
+  Go, "go" => &["go", "golang"],
+  Haskell, "haskell" => &["hs", "haskell"],
+  Html, "html" => &["html"],
+  Java, "java" => &["java"],
+  JavaScript, "javascript" => &["javascript", "js", "jsx"],
+  Json, "json" => &["json"],
+  Kotlin, "kotlin" => &["kotlin", "kt"],
+  Lua, "lua" => &["lua"],
+  Php, "php" => &["php"],
+  Python, "python" => &["py", "python"],
+  Ruby, "ruby" => &["rb", "ruby"],
+  Rust, "rust" => &["rs", "rust"],
+  Scala, "scala" => &["scala"],
+  Swift, "swift" => &["swift"],
+  TypeScript, "typescript" => &["ts", "typescript"],
+  Tsx, "tsx" => &["tsx"],
+  Yaml, "yaml" => &["yaml", "yml"],
 }
 
 /// Implements the language names and aliases.
@@ -421,28 +676,51 @@ impl FromStr for SupportLang {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         // Fast path: try exact matches first (most common case)
         match s {
+            #[cfg(feature = "bash")]
             "bash" => return Ok(SupportLang::Bash),
+            #[cfg(feature = "c")]
             "c" => return Ok(SupportLang::C),
+            #[cfg(feature = "cpp")]
             "cpp" | "c++" => return Ok(SupportLang::Cpp),
+            #[cfg(feature = "csharp")]
             "cs" | "csharp" => return Ok(SupportLang::CSharp),
+            #[cfg(feature = "css")]
             "css" => return Ok(SupportLang::Css),
+            #[cfg(feature = "elixir")]
             "elixir" | "ex" => return Ok(SupportLang::Elixir),
+            #[cfg(feature = "go")]
             "go" | "golang" => return Ok(SupportLang::Go),
+            #[cfg(feature = "haskell")]
             "haskell" | "hs" => return Ok(SupportLang::Haskell),
+            #[cfg(feature = "html")]
             "html" => return Ok(SupportLang::Html),
+            #[cfg(feature = "java")]
             "java" => return Ok(SupportLang::Java),
+            #[cfg(feature = "javascript")]
             "javascript" | "js" => return Ok(SupportLang::JavaScript),
+            #[cfg(feature = "json")]
             "json" => return Ok(SupportLang::Json),
+            #[cfg(feature = "kotlin")]
             "kotlin" | "kt" => return Ok(SupportLang::Kotlin),
+            #[cfg(feature = "lua")]
             "lua" => return Ok(SupportLang::Lua),
+            #[cfg(feature = "php")]
             "php" => return Ok(SupportLang::Php),
+            #[cfg(feature = "python")]
             "python" | "py" => return Ok(SupportLang::Python),
+            #[cfg(feature = "ruby")]
             "ruby" | "rb" => return Ok(SupportLang::Ruby),
+            #[cfg(feature = "rust")]
             "rust" | "rs" => return Ok(SupportLang::Rust),
+            #[cfg(feature = "scala")]
             "scala" => return Ok(SupportLang::Scala),
+            #[cfg(feature = "swift")]
             "swift" => return Ok(SupportLang::Swift),
+            #[cfg(feature = "typescript")]
             "typescript" | "ts" => return Ok(SupportLang::TypeScript),
+            #[cfg(feature = "tsx")]
             "tsx" => return Ok(SupportLang::Tsx),
+            #[cfg(feature = "yaml")]
             "yaml" | "yml" => return Ok(SupportLang::Yaml),
             _ => {} // Fall through to case-insensitive search
         }
@@ -463,28 +741,51 @@ macro_rules! execute_lang_method {
   ($me: path, $method: ident, $($pname:tt),*) => {
     use SupportLang as S;
     match $me {
+      #[cfg(feature = "bash")]
       S::Bash => Bash.$method($($pname,)*),
+      #[cfg(feature = "c")]
       S::C => C.$method($($pname,)*),
+        #[cfg(feature = "cpp")]
       S::Cpp => Cpp.$method($($pname,)*),
+        #[cfg(feature = "csharp")]
       S::CSharp => CSharp.$method($($pname,)*),
+        #[cfg(feature = "css")]
       S::Css => Css.$method($($pname,)*),
+        #[cfg(feature = "elixir")]
       S::Elixir => Elixir.$method($($pname,)*),
+        #[cfg(feature = "go")]
       S::Go => Go.$method($($pname,)*),
+        #[cfg(feature = "haskell")]
       S::Haskell => Haskell.$method($($pname,)*),
+        #[cfg(feature = "html")]
       S::Html => Html.$method($($pname,)*),
+        #[cfg(feature = "json")]
       S::Java => Java.$method($($pname,)*),
+        #[cfg(feature = "javascript")]
       S::JavaScript => JavaScript.$method($($pname,)*),
+        #[cfg(feature = "json")]
       S::Json => Json.$method($($pname,)*),
+        #[cfg(feature = "kotlin")]
       S::Kotlin => Kotlin.$method($($pname,)*),
+        #[cfg(feature = "lua")]
       S::Lua => Lua.$method($($pname,)*),
+        #[cfg(feature = "php")]
       S::Php => Php.$method($($pname,)*),
+        #[cfg(feature = "python")]
       S::Python => Python.$method($($pname,)*),
+        #[cfg(feature = "ruby")]
       S::Ruby => Ruby.$method($($pname,)*),
+        #[cfg(feature = "rust")]
       S::Rust => Rust.$method($($pname,)*),
+        #[cfg(feature = "scala")]
       S::Scala => Scala.$method($($pname,)*),
+        #[cfg(feature = "swift")]
       S::Swift => Swift.$method($($pname,)*),
+        #[cfg(feature = "typescript")]
       S::Tsx => Tsx.$method($($pname,)*),
+        #[cfg(feature = "typescript")]
       S::TypeScript => TypeScript.$method($($pname,)*),
+        #[cfg(feature = "yaml")]
       S::Yaml => Yaml.$method($($pname,)*),
     }
   }
@@ -504,6 +805,7 @@ impl Language for SupportLang {
     impl_lang_method!(meta_var_char, () => char);
     impl_lang_method!(expando_char, () => char);
     impl_lang_method!(extract_meta_var, (source: &str) => Option<MetaVariable>);
+    #[cfg(feature = "matching")]
     impl_lang_method!(build_pattern, (builder: &PatternBuilder) => Result<Pattern, PatternError>);
     fn pre_process_pattern<'q>(&self, query: &'q str) -> Cow<'q, str> {
         execute_lang_method! { self, pre_process_pattern, query }
@@ -513,6 +815,7 @@ impl Language for SupportLang {
     }
 }
 
+#[cfg(feature = "matching")]
 impl LanguageExt for SupportLang {
     impl_lang_method!(get_ts_language, () => TSLanguage);
     impl_lang_method!(injectable_languages, () => Option<&'static [&'static str]>);
@@ -521,6 +824,7 @@ impl LanguageExt for SupportLang {
         root: Node<StrDoc<L>>,
     ) -> RapidMap<String, Vec<TSRange>> {
         match self {
+            #[cfg(feature = "html-embedded")]
             SupportLang::Html => Html.extract_injections(root),
             _ => RapidMap::default(),
         }
@@ -530,30 +834,53 @@ impl LanguageExt for SupportLang {
 const fn extensions(lang: SupportLang) -> &'static [&'static str] {
     use SupportLang::*;
     match lang {
+        #[cfg(feature = "bash")]
         Bash => &[
             "bash", "bats", "cgi", "command", "env", "fcgi", "ksh", "sh", "tmux", "tool", "zsh",
         ],
+        #[cfg(feature = "c")]
         C => &["c", "h"],
+        #[cfg(feature = "cpp")]
         Cpp => &["cc", "hpp", "cpp", "c++", "hh", "cxx", "cu", "ino"],
+        #[cfg(feature = "csharp")]
         CSharp => &["cs"],
+        #[cfg(feature = "css")]
         Css => &["css", "scss"],
+        #[cfg(feature = "elixir")]
         Elixir => &["ex", "exs"],
+        #[cfg(feature = "go")]
         Go => &["go"],
+        #[cfg(feature = "haskell")]
         Haskell => &["hs"],
+        #[cfg(feature = "html")]
         Html => &["html", "htm", "xhtml"],
+        #[cfg(feature = "java")]
         Java => &["java"],
+        #[cfg(feature = "javascript")]
         JavaScript => &["cjs", "js", "mjs", "jsx"],
+        #[cfg(feature = "json")]
         Json => &["json"],
+        #[cfg(feature = "kotlin")]
         Kotlin => &["kt", "ktm", "kts"],
+        #[cfg(feature = "lua")]
         Lua => &["lua"],
+        #[cfg(feature = "php")]
         Php => &["php"],
+        #[cfg(feature = "python")]
         Python => &["py", "py3", "pyi", "bzl"],
+        #[cfg(feature = "ruby")]
         Ruby => &["rb", "rbw", "gemspec"],
+        #[cfg(feature = "rust")]
         Rust => &["rs"],
+        #[cfg(feature = "scala")]
         Scala => &["scala", "sc", "sbt"],
+        #[cfg(feature = "swift")]
         Swift => &["swift"],
+        #[cfg(feature = "typescript")]
         TypeScript => &["ts", "cts", "mts"],
+        #[cfg(feature = "tsx")]
         Tsx => &["tsx"],
+        #[cfg(feature = "yaml")]
         Yaml => &["yaml", "yml"],
     }
 }
@@ -566,18 +893,35 @@ fn from_extension(path: &Path) -> Option<SupportLang> {
 
     // Fast path: try most common extensions first
     match ext {
-        "rs" => return Some(SupportLang::Rust),
-        "js" | "mjs" | "cjs" => return Some(SupportLang::JavaScript),
-        "ts" | "cts" | "mts" => return Some(SupportLang::TypeScript),
-        "tsx" => return Some(SupportLang::Tsx),
-        "py" | "py3" | "pyi" => return Some(SupportLang::Python),
-        "java" => return Some(SupportLang::Java),
+        #[cfg(feature = "bash")]
+        "bash" | "sh" | ".bashrc" | "bash_aliases" | "bats" | "cgi" | "command" | "env" | "fcgi" | "ksh" | "tmux" | "tool" | "zsh" | "bash_logout" | "bash_profile" | "profile" | "login" | "logout" => {
+            return Some(SupportLang::Bash)
+        }
+        #[cfg(feature = "c")]
+        "c" | "h" => return Some(SupportLang::C),
+        #[cfg(feature = "csharp")]
         "cpp" | "cc" | "cxx" => return Some(SupportLang::Cpp),
-        "c" => return Some(SupportLang::C),
-        "go" => return Some(SupportLang::Go),
-        "html" | "htm" => return Some(SupportLang::Html),
+        #[cfg(feature = "css")]
         "css" => return Some(SupportLang::Css),
+        #[cfg(feature = "go")]
+        "go" => return Some(SupportLang::Go),
+        #[cfg(feature = "html")]
+        "html" | "htm" => return Some(SupportLang::Html),
+        #[cfg(feature = "java")]
+        "java" => return Some(SupportLang::Java),
+        #[cfg(feature = "javascript")]
+        "js" | "mjs" | "cjs" => return Some(SupportLang::JavaScript),
+        #[cfg(feature = "json")]
         "json" => return Some(SupportLang::Json),
+        #[cfg(feature = "python")]
+        "py" | "py3" | "pyi" => return Some(SupportLang::Python),
+        #[cfg(feature = "rust")]
+        "rs" => return Some(SupportLang::Rust),
+        #[cfg(feature = "typescript")]
+        "ts" | "cts" | "mts" => return Some(SupportLang::TypeScript),
+        #[cfg(feature = "tsx")]
+        "tsx" => return Some(SupportLang::Tsx),
+        #[cfg(feature = "yaml")]
         "yaml" | "yml" => return Some(SupportLang::Yaml),
         _ => {}
     }
